feat: adds user auth validation

+++ src/Tasks.Application/Services/UserService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
+using System.Reflection;
 using System.Security.Claims;
 using Tasks.Application.DTOs.User;
 using Tasks.Application.DTOs.User.Validators;
@@ -23,11 +24,25 @@ public class UserService : IUserService
         _httpContext = httpContext;
     }
 
-    public async Task<ResultViewModel<List<User>>> GetAll()
-        =>  new ResultViewModel<List<User>>(await _userRepository.GetAll());
+    public async Task<ResultViewModel<List<User>>> GetAll() 
+    {
+        if (!_httpContext.HttpContext.User.IsInRole("Admin"))
+        {
+            _httpContext.HttpContext.Response.StatusCode = StatusCodes.Status403Forbidden;
+            return new ResultViewModel<List<User>>("Acesso não autorizado");
+        }
+
+        return new ResultViewModel<List<User>>(await _userRepository.GetAll()); 
+    }
 
     public async Task<ResultViewModel<User>> GetById(Guid id)
     {
+        if (!_httpContext.HttpContext.User.IsInRole("Admin") && id != GetUserIdFromJwt())
+        {
+            _httpContext.HttpContext.Response.StatusCode = StatusCodes.Status403Forbidden;
+            return new ResultViewModel<User>("Acesso não autorizado");
+        }
+
         var user = await _userRepository.GetById(id);
 
         if (user is null)
@@ -45,10 +60,10 @@ public class UserService : IUserService
         }
 
         var validationResult = await new CreateMissionDTOValidator().ValidateAsync(missionDto);
-        
+
         if (!validationResult.IsValid)
             return new ResultViewModel<User>(validationResult);
-        
+
         var user = await _userRepository.GetById(missionDto.UserId);
 
         if (user is null)
@@ -56,7 +71,7 @@ public class UserService : IUserService
 
         var mission = new Mission(missionDto.Title, missionDto.Description, missionDto.UserId);
         user.AddMission(mission);
-        
+
         _missionRepository.Create(user.Missions.FirstOrDefault(m => m.Id == mission.Id));
 
         return new ResultViewModel<User>(user);
@@ -65,10 +80,10 @@ public class UserService : IUserService
     public async Task<ResultViewModel<User>> UpdateMissionsPosition(EditMissionsPositionDTO missionDto)
     {
         var validationResult = await new EditMissionsPositionDTOValidator().ValidateAsync(missionDto);
-        
+
         if (!validationResult.IsValid)
             return new ResultViewModel<User>(validationResult);
-        
+
         var mission = await _missionRepository.GetById(missionDto.MissionId);
 
         if (mission is null)
@@ -83,7 +98,7 @@ public class UserService : IUserService
         var user = await _userRepository.GetById(mission.UserId);
 
         user.UpdateMissionPosition(missionDto.MissionId, missionDto.NewPosition);
-        
+
         _missionRepository.UpdateMany(user.Missions);
 
         return new ResultViewModel<User>(user);
@@ -92,10 +107,10 @@ public class UserService : IUserService
     public async Task<ResultViewModel<Mission>> UpdateMission(EditMissionDTO missionDto)
     {
         var validationResult = await new EditMissionDTOValidator().ValidateAsync(missionDto);
-        
+
         if (!validationResult.IsValid)
             return new ResultViewModel<Mission>(validationResult);
-        
+
         var mission = await _missionRepository.GetById(missionDto.Id);
 
         if (!_httpContext.HttpContext.User.IsInRole("Admin") && mission.UserId != GetUserIdFromJwt())
@@ -107,7 +122,7 @@ public class UserService : IUserService
         mission.Title = missionDto.Title;
         mission.Description = missionDto.Description;
         mission.Status = missionDto.Status;
-        
+
         _missionRepository.Update(mission);
 
         return new ResultViewModel<Mission>(mission);
@@ -116,14 +131,20 @@ public class UserService : IUserService
     public async Task<ResultViewModel<User>> UpdateUserInfo(EditUserDTO userDto)
     {
         var validationResult = await new EditUserDTOValidator().ValidateAsync(userDto);
-        
+
         if (!validationResult.IsValid)
             return new ResultViewModel<User>(validationResult);
+
+        if (!_httpContext.HttpContext.User.IsInRole("Admin") && userDto.Id != GetUserIdFromJwt())
+        {
+            _httpContext.HttpContext.Response.StatusCode = StatusCodes.Status403Forbidden;
+            return new ResultViewModel<User>("Acesso não autorizado");
+        }
 
         var user = new User(userDto.Name, userDto.Email);
 
         user.Id = userDto.Id;
-        
+
         _userRepository.Update(user);
 
         return new ResultViewModel<User>(user);
@@ -133,9 +154,15 @@ public class UserService : IUserService
     {
         var user = await _userRepository.GetById(id);
 
+        if (!_httpContext.HttpContext.User.IsInRole("Admin") && id != GetUserIdFromJwt())
+        {
+            _httpContext.HttpContext.Response.StatusCode = StatusCodes.Status403Forbidden;
+            return new ResultViewModel<User>("Acesso não autorizado");
+        }
+
         if (user is null)
             return new ResultViewModel<User>("User not found");
-        
+
         _userRepository.Delete(id);
 
         return new ResultViewModel<User>(user);
@@ -157,9 +184,9 @@ public class UserService : IUserService
         var user = await _userRepository.GetById(mission.UserId);
 
         user.RemoveMission(missionId);
-        
+
         _missionRepository.Delete(missionId);
-        
+
         _missionRepository.UpdateMany(user.Missions);
 
         return new ResultViewModel<User>(user);
